+++ app/code/community/Jtl/Connector/src/Mapper/Order.php
@@ -28,7 +28,8 @@ class Order
 {
     public static $paymentMethods = array(
         'amazonpayments_advanced' => PaymentTypes::TYPE_AMAPAY,
-        'bankpayment' => PaymentTypes::TYPE_BANK_TRANSFER,
+        'bankpayment' => PaymentTypes::TYPE_PREPAYMENT,
+        'banktransfer' => PaymentTypes::TYPE_BANK_TRANSFER,
         'cash' => PaymentTypes::TYPE_CASH,
         'cashondelivery' => PaymentTypes::TYPE_CASH_ON_DELIVERY,
         'checkmo' => PaymentTypes::TYPE_BANK_TRANSFER,
@@ -46,6 +47,10 @@ class Order
         'paypaluk_express' => PaymentTypes::TYPE_PAYPAL_EXPRESS,
         'phoenix_cashondelivery' => PaymentTypes::TYPE_CASH_ON_DELIVERY,
         'saferpaynew' => PaymentTypes::TYPE_SAFERPAY
+    );
+
+    public static $postpaidMethods = array(
+        PaymentTypes::TYPE_BANK_TRANSFER
     );
 
     public function getAvailableCount()
@@ -133,14 +138,19 @@ class Order
             $payment = $order->getPayment();
             $code = $payment->getMethodInstance()->getCode();
 
-            if (in_array($order->getState(), array(\Mage_Sales_Model_Order::STATE_PROCESSING, \Mage_Sales_Model_Order::STATE_COMPLETE))) {
-                $customerOrder->setPaymentStatus(ConnectorCustomerOrder::PAYMENT_STATUS_COMPLETED);
-            }
-
             if (array_key_exists($code, self::$paymentMethods))
                 $customerOrder->setPaymentModuleCode(self::$paymentMethods[$code]);
             else
-                $customerOrder->setPaymentModuleCode('pm_bank_transfer');
+                $customerOrder->setPaymentModuleCode(PaymentTypes::TYPE_PREPAYMENT);
+
+            if (in_array($order->getState(), array(\Mage_Sales_Model_Order::STATE_PROCESSING, \Mage_Sales_Model_Order::STATE_COMPLETE))) {
+                if (!in_array($customerOrder->getPaymentModuleCode(), self::$postpaidMethods)) {
+                    $customerOrder->setPaymentStatus(ConnectorCustomerOrder::PAYMENT_STATUS_COMPLETED);
+                }
+                else {
+                    $customerOrder->setPaymentStatus(ConnectorCustomerOrder::PAYMENT_STATUS_UNPAID);
+                }
+            }
 
             foreach ($order->getAllItems() as $magento_item) {
                 $item = new ConnectorCustomerOrderItem();