+++ app/code/community/Jtl/Connector/src/Mapper/Specific.php
@@ -91,6 +91,13 @@ class Specific
             $frontendLabels[$storeId] = $specificI18n->getName();
         }
 
+        if ($specific->getType() === 'SELECTBOX') {
+            $frontendType = 'multiselect';
+        }
+        else {
+            $frontendType = 'textarea';
+        }
+
         Logger::write('Creating specific: ' . $attributeCode, Logger::DEBUG);
         $attributeData = array(
             'attribute_code' => $attributeCode,
@@ -106,7 +113,7 @@ class Specific
             'used_in_product_listing' => 1,
             'used_for_sort_by' => 1,
             'is_configurable' => 0,
-            'frontend_input' => 'multiselect',
+            'frontend_input' => $frontendType,
             'is_wysiwyg_enabled' => 0,
             'is_unique' => 0,
             'is_required' => 0,
@@ -260,50 +267,7 @@ class Specific
             'value' => $options
         ));
         $attrModel->setData('default_value', $defaultValue);
-//        die(var_dump($attrModel));
         $attrModel->save();
-
-//        foreach ($specific->getValues() as $specificValue) {
-//            $defaultSpecificValueI18n = ArrayTools::filterOneByLanguage($specificValue->getI18ns(), $defaultLanguageIso);
-//            if ($defaultSpecificValueI18n === null)
-//                $defaultSpecificValueI18n = reset($specificValue->getI18ns());
-//            $matches = array_filter($values, function ($value) use ($defaultSpecificValueI18n) {
-//                return ($value['label'] === $defaultSpecificValueI18n->getValue());
-//            });
-//
-//            // Value found
-//            if ($matches)
-//                continue;
-//
-//            Logger::write(sprintf('value "%s" not found', $specificValue->getId()->getHost()), Logger::DEBUG);
-//
-//            $attribute_model = \Mage::getModel('eav/entity_attribute');
-//            $attribute_options_model = \Mage::getModel('eav/entity_attribute_source_table');
-//
-//            $attribute_table = $attribute_options_model->setAttribute($attribute);
-//            $options = $attribute_options_model->getAllOptions(false);
-//            Logger::write(var_export($options, true), Logger::DEBUG);
-//
-//            $stores = Magento::getInstance()->getStoreMapping();
-//            $newAttributeValue = array('option' => array());
-//            $newAttributeValue['option'] = array(
-//                \Mage_Core_Model_App::ADMIN_STORE_ID => $defaultSpecificValueI18n->getValue()
-//            );
-//            foreach ($stores as $locale => $storeId) {
-//                $specificValueI18n = ArrayTools::filterOneByLanguage($specificValue->getI18ns(), LocaleMapper::localeToLanguageIso($locale));
-//                if ($specificValueI18n === null) {
-//                    $i18ns = $specificValue->getI18ns();
-//                    $specificValueI18n = reset($i18ns);
-//                }
-//
-//                $newAttributeValue['option'][$storeId] = $specificValueI18n->getValue();
-//            }
-//            $result = array('value' => $newAttributeValue);
-//            $attribute->setData('option', $result);
-//            die(var_dump($attribute));
-//            $attribute->save();
-//
-//        }
     }
 
     public function update(ConnectorSpecific $specific)