+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 --------------------
 - Update jtl/connector to version 2.2.4
 - Import specifics as select or textarea depending on the type selected in JTL-Wawi
+- Fix payment status for postpaid methods
 
 1.4.2.0
 -------