+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 1.4.3.0 (unreleased)
 --------------------
-- Update jtl/connector to version 2.2.4
+- Update jtl/connector to version 2.2.5
 - Import specifics as select or textarea depending on the type selected in JTL-Wawi
 - Fix payment status for postpaid methods